+++ lib/src/functions.dart
@@ -870,24 +870,9 @@ void defineCoreFunctions(Environment environment) {
     return new SassBoolean(_features.contains(feature.text));
   });
 
-  environment.defineFunction("variable-exists", r"$name", (arguments) {
-    var variable = arguments[0].assertString("name");
-    return new SassBoolean(environment.variableExists(variable.text));
-  });
-
   environment.defineFunction("global-variable-exists", r"$name", (arguments) {
     var variable = arguments[0].assertString("name");
     return new SassBoolean(environment.globalVariableExists(variable.text));
-  });
-
-  environment.defineFunction("function-exists", r"$name", (arguments) {
-    var variable = arguments[0].assertString("name");
-    return new SassBoolean(environment.functionExists(variable.text));
-  });
-
-  environment.defineFunction("mixin-exists", r"$name", (arguments) {
-    var variable = arguments[0].assertString("name");
-    return new SassBoolean(environment.mixinExists(variable.text));
   });
 
   environment.defineFunction("inspect", r"$value",